requirements.txt
# Core Dependencies
langchain>=0.1.20
langgraph>=0.0.60
langchain-openai>=0.1.0
langchain-community>=0.0.38

# Web Interface
gradio==4.15.0
fastapi==0.109.0
uvicorn==0.27.0

# Database & Storage
sqlalchemy==2.0.25
alembic==1.13.1

# Text-to-Speech
gtts==2.5.1
pydub==0.25.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
typing-extensions==4.9.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0

# Optional: For enhanced capabilities
pillow==10.2.0
matplotlib==3.8.2
plotly==5.18.0